+++ browser-test/src/admin/admin_application_statuses.test.ts
@@ -13,6 +13,7 @@ import {
   extractEmailsForRecipient,
   validateScreenshot,
   AdminProgramStatuses,
+  enableFeatureFlag,
 } from '../support'
 
 test.describe('view program statuses', () => {
@@ -25,6 +26,9 @@ test.describe('view program statuses', () => {
     const programWithoutStatusesName = 'Test program without statuses'
     test.beforeEach(async ({page, adminPrograms, applicantQuestions}) => {
       await loginAsAdmin(page)
+
+      // enable bulk status feature flag
+      await enableFeatureFlag(page, 'bulk_status_update_enabled')
 
       // Add a program, no questions are needed.
       await adminPrograms.addProgram(programWithoutStatusesName)
@@ -41,42 +45,49 @@ test.describe('view program statuses', () => {
       // Navigate to the submitted application as the program admin.
       await loginAsProgramAdmin(page)
       await adminPrograms.viewApplications(programWithoutStatusesName)
-      await adminPrograms.viewApplicationForApplicant('Guest')
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
     })
 
     test('does not show status options', async ({adminPrograms}) => {
-      expect(await adminPrograms.isStatusSelectorVisible()).toBe(false)
+      expect(await adminPrograms.isStatusSelectorVisibleForBulkStatus()).toBe(
+        false,
+      )
     })
 
-    test('does not show application status in list', async ({
+    test('does not show application status in table', async ({
+      page,
       adminPrograms,
     }) => {
-      await adminPrograms.expectApplicationStatusDoesntContain(
+      await page.getByRole('link', {name: 'Back'}).click()
+      await adminPrograms.expectApplicationStatusDoesntContainForBulkStatus(
         'Guest',
         'Status: ',
       )
     })
 
-    test('does not show edit note', async ({adminPrograms}) => {
-      expect(await adminPrograms.isEditNoteVisible()).toBe(false)
+    test('does not show edit note', async ({page, adminPrograms}) => {
+      expect(await adminPrograms.isEditNoteVisibleForBulkStatus()).toBe(false)
+      await page.getByRole('link', {name: 'Back'}).click()
     })
 
     test('does not show pagination when there is only 1 page of applications', async ({
       adminPrograms,
     }) => {
-      expect(await adminPrograms.isPaginationVisibleForApplicationList()).toBe(
+      expect(await adminPrograms.isPaginationVisibleForApplicationTable()).toBe(
         false,
       )
     })
 
     /* See trusted_intermediary.test.ts for more comprehensive pagination testing */
-    test('shows pagination if there are more than 10 applications', async ({
+    test.skip('shows pagination if there are more than 100 applications', async ({
       page,
       adminPrograms,
       applicantQuestions,
     }) => {
-      // There is already 1 application from the beforeEach, so apply to 10 more programs.
-      for (let i = 0; i < 10; i++) {
+      test.slow()
+      await page.getByRole('link', {name: 'Back'}).click()
+      // There is already 1 application from the beforeEach, so apply to 100 more programs.
+      for (let i = 0; i < 100; i++) {
         await logout(page)
 
         // Submit an application as a guest.
@@ -85,18 +96,17 @@ test.describe('view program statuses', () => {
         )
         await applicantQuestions.submitFromReviewPage()
 
-        // Navigate to the submitted application as the program admin.
         await loginAsProgramAdmin(page)
       }
 
       // Navigate to the applications list
       await adminPrograms.viewApplications(programWithoutStatusesName)
 
-      await validateScreenshot(page, 'application-list-pagination')
-      expect(await adminPrograms.isPaginationVisibleForApplicationList()).toBe(
+      expect(await adminPrograms.isPaginationVisibleForApplicationTable()).toBe(
         true,
       )
       expect(page.locator('.usa-pagination__button:has-text("2")'))
+      await validateScreenshot(page, 'application-table-pagination')
     })
   })
 
@@ -114,132 +124,172 @@ test.describe('view program statuses', () => {
           applicantQuestions,
           adminProgramStatuses,
         )
+        // enable bulk status feature flag
+        await enableFeatureFlag(page, 'bulk_status_update_enabled')
         await loginAsProgramAdmin(page)
         await adminPrograms.viewApplications(programWithStatusesName)
-        await adminPrograms.viewApplicationForApplicant('Guest')
+        await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
       },
     )
 
     test('shows status selector', async ({adminPrograms}) => {
-      expect(await adminPrograms.isStatusSelectorVisible()).toBe(true)
+      expect(await adminPrograms.isStatusSelectorVisibleForBulkStatus()).toBe(
+        true,
+      )
     })
 
     test('shows placeholder option', async ({adminPrograms}) => {
-      expect(await adminPrograms.getStatusOption()).toBe('Choose an option:')
+      expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+        'Choose an option:',
+      )
     })
 
     test('renders', async ({page}) => {
       await validateScreenshot(page, 'application-view-with-statuses')
     })
 
-    test('shows "None" value in application list if no status is set', async ({
+    test('shows "None" value in application table if no status is set', async ({
       adminPrograms,
     }) => {
       await adminPrograms.viewApplications(programWithStatusesName)
-      await adminPrograms.expectApplicationHasStatusString('Guest', 'None')
+      await adminPrograms.expectApplicationHasStatusStringForBulkStatus(
+        'Guest',
+        'None',
+      )
     })
 
     test.describe('when a status is changed, a confirmation dialog is shown', () => {
       test('renders', async ({page, adminPrograms}) => {
-        await adminPrograms.setStatusOptionAndAwaitModal(noEmailStatusName)
+        await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+          noEmailStatusName,
+        )
         await validateScreenshot(page, 'change-status-modal')
       })
 
       test('when rejecting, the selected status is not changed', async ({
+        page,
         adminPrograms,
       }) => {
-        await adminPrograms.setStatusOptionAndAwaitModal(noEmailStatusName)
-        await dismissModal(adminPrograms.applicationFrame())
-        expect(await adminPrograms.getStatusOption()).toBe('Choose an option:')
+        await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+          noEmailStatusName,
+        )
+        await dismissModal(page)
+        expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+          'Choose an option:',
+        )
       })
 
-      test('when confirmed, the page is redirected with a success toast and preserves the selected application', async ({
+      test('when confirmed, the page is shown a success toast', async ({
         adminPrograms,
       }) => {
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(noEmailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            noEmailStatusName,
+          )
         expect(await modal.innerText()).toContain(
           `Status Change: Unset -> ${noEmailStatusName}`,
         )
-        await adminPrograms.confirmStatusUpdateModal(modal)
-        expect(await adminPrograms.getStatusOption()).toBe(noEmailStatusName)
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+        expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+          noEmailStatusName,
+        )
         await adminPrograms.expectUpdateStatusToast()
-
-        // Confirm that the application is shown after reloading the page.
-        const applicationText = await adminPrograms
-          .applicationFrameLocator()
-          .locator('#application-view')
-          .innerText()
-        expect(applicationText).toContain('Guest')
       })
 
       test('when no email is configured for the status, a warning is shown', async ({
+        page,
         adminPrograms,
       }) => {
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(noEmailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            noEmailStatusName,
+          )
         expect(await modal.innerText()).toContain(
           'will not receive an email because there is no email content set for this status. Connect with your CiviForm Admin to add an email to this status',
         )
-        await dismissModal(adminPrograms.applicationFrame())
+        await dismissModal(page)
       })
 
       test('when no email is configured for the applicant, a warning is shown', async ({
+        page,
         adminPrograms,
       }) => {
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            emailStatusName,
+          )
         expect(await modal.innerText()).toContain(
           'will not receive an email for this change since they have not provided an email address.',
         )
-        await dismissModal(adminPrograms.applicationFrame())
+        await dismissModal(page)
       })
 
       test('when changing status, the previous status is shown', async ({
+        page,
         adminPrograms,
       }) => {
         await test.step('Set initial status', async () => {
           const modal =
-            await adminPrograms.setStatusOptionAndAwaitModal(noEmailStatusName)
-          await adminPrograms.confirmStatusUpdateModal(modal)
-          expect(await adminPrograms.getStatusOption()).toBe(noEmailStatusName)
+            await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+              noEmailStatusName,
+            )
+          await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+          expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+            noEmailStatusName,
+          )
         })
 
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            emailStatusName,
+          )
         expect(await modal.innerText()).toContain(
           `Status Change: ${noEmailStatusName} -> ${emailStatusName}`,
         )
-        await dismissModal(adminPrograms.applicationFrame())
+        await dismissModal(page)
       })
 
-      test('when changing status, the updated application status is reflected in the application list', async ({
+      test('when changing status, the updated application status is reflected in the application table', async ({
+        page,
         adminPrograms,
       }) => {
         await test.step('Set initial status', async () => {
           const modal =
-            await adminPrograms.setStatusOptionAndAwaitModal(noEmailStatusName)
-          await adminPrograms.confirmStatusUpdateModal(modal)
-          expect(await adminPrograms.getStatusOption()).toBe(noEmailStatusName)
+            await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+              noEmailStatusName,
+            )
+          await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+          expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+            noEmailStatusName,
+          )
         })
 
-        await adminPrograms.expectApplicationHasStatusString(
+        await page.getByRole('link', {name: 'Back'}).click()
+
+        await adminPrograms.expectApplicationHasStatusStringForBulkStatus(
           'Guest',
           noEmailStatusName,
         )
+        await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
-        await adminPrograms.confirmStatusUpdateModal(modal)
-        await adminPrograms.expectApplicationHasStatusString(
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            emailStatusName,
+          )
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+        await page.getByRole('link', {name: 'Back'}).click()
+        await adminPrograms.expectApplicationHasStatusStringForBulkStatus(
           'Guest',
           emailStatusName,
         )
       })
 
       test.describe('when email is configured for the status and applicant, a checkbox is shown to notify the applicant', () => {
-        test.beforeEach(async ({adminPrograms}) => {
+        test.beforeEach(async ({page, adminPrograms}) => {
+          await enableFeatureFlag(page, 'bulk_status_update_enabled')
           await adminPrograms.viewApplications(programWithStatusesName)
-          await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
+          await adminPrograms.viewApplicationForApplicantForBulkStatus(
+            testUserDisplayName(),
+          )
         })
 
         test('choosing not to notify applicant changes status and does not send email', async ({
@@ -250,13 +300,17 @@ test.describe('view program statuses', () => {
             ? await extractEmailsForRecipient(page, testUserDisplayName())
             : []
           const modal =
-            await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+            await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+              emailStatusName,
+            )
           const notifyCheckbox = await modal.$('input[type=checkbox]')
           expect(notifyCheckbox).not.toBeNull()
           await notifyCheckbox!.uncheck()
           expect(await notifyCheckbox!.isChecked()).toBe(false)
-          await adminPrograms.confirmStatusUpdateModal(modal)
-          expect(await adminPrograms.getStatusOption()).toBe(emailStatusName)
+          await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+          expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+            emailStatusName,
+          )
           await adminPrograms.expectUpdateStatusToast()
 
           if (supportsEmailInspection()) {
@@ -276,13 +330,17 @@ test.describe('view program statuses', () => {
             ? await extractEmailsForRecipient(page, testUserDisplayName())
             : []
           const modal =
-            await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+            await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+              emailStatusName,
+            )
           const notifyCheckbox = await modal.$('input[type=checkbox]')
           expect(notifyCheckbox).not.toBeNull()
           expect(await notifyCheckbox!.isChecked()).toBe(true)
           expect(await modal.innerText()).toContain(' of this change at ')
-          await adminPrograms.confirmStatusUpdateModal(modal)
-          expect(await adminPrograms.getStatusOption()).toBe(emailStatusName)
+          await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+          expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+            emailStatusName,
+          )
           await adminPrograms.expectUpdateStatusToast()
 
           if (supportsEmailInspection()) {
@@ -298,21 +356,21 @@ test.describe('view program statuses', () => {
     })
 
     test('allows editing a note and preserves the selected application', async ({
+      page,
       adminPrograms,
     }) => {
-      await adminPrograms.editNote('Some note content')
+      await adminPrograms.editNoteForBulkStatus('Some note content')
       await adminPrograms.expectNoteUpdatedToast()
 
       // Confirm that the application is shown after reloading the page.
-      const applicationText = await adminPrograms
-        .applicationFrameLocator()
+      const applicationText = await page
         .locator('#application-view')
         .innerText()
       expect(applicationText).toContain('Guest')
     })
 
     test('renders the note dialog', async ({page, adminPrograms}) => {
-      await adminPrograms.awaitEditNoteModal()
+      await adminPrograms.awaitEditNoteModalForBulkStatus()
       await page.evaluate(() => {
         window.scrollTo(0, 0)
       })
@@ -321,66 +379,66 @@ test.describe('view program statuses', () => {
 
     test('shows the current note content', async ({adminPrograms}) => {
       const noteText = 'Some note content'
-      await adminPrograms.editNote(noteText)
+      await adminPrograms.editNoteForBulkStatus(noteText)
       await adminPrograms.expectNoteUpdatedToast()
 
       // Reload the note editor.
       await adminPrograms.viewApplications(programWithStatusesName)
-      await adminPrograms.viewApplicationForApplicant('Guest')
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
 
-      expect(await adminPrograms.getNoteContent()).toBe(noteText)
+      expect(await adminPrograms.getNoteContentForBulkStatus()).toBe(noteText)
     })
 
     test('allows updating a note', async ({adminPrograms}) => {
       const noteText = 'Some note content'
-      await adminPrograms.editNote('first note content')
+      await adminPrograms.editNoteForBulkStatus('first note content')
       await adminPrograms.expectNoteUpdatedToast()
-      await adminPrograms.editNote(noteText)
+      await adminPrograms.editNoteForBulkStatus(noteText)
       await adminPrograms.expectNoteUpdatedToast()
 
       // Reload the note editor.
       await adminPrograms.viewApplications(programWithStatusesName)
-      await adminPrograms.viewApplicationForApplicant('Guest')
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
 
-      expect(await adminPrograms.getNoteContent()).toBe(noteText)
+      expect(await adminPrograms.getNoteContentForBulkStatus()).toBe(noteText)
     })
-    test('allow notes to be exported', async ({adminPrograms}) => {
-      await adminPrograms.editNote('Note is exported')
+    test('allow notes to be exported', async ({page, adminPrograms}) => {
+      await adminPrograms.editNoteForBulkStatus('Note is exported')
       await adminPrograms.expectNoteUpdatedToast()
       const noApplyFilters = false
+      await page.getByRole('link', {name: 'Back'}).click()
 
       const csvContent = await adminPrograms.getCsv(noApplyFilters)
       expect(csvContent).toContain('Note is exported')
     })
 
-    test('export only the latest note', async ({adminPrograms}) => {
-      await adminPrograms.editNote('Note is exported')
+    test('export only the latest note', async ({page, adminPrograms}) => {
+      await adminPrograms.editNoteForBulkStatus('Note is exported')
       await adminPrograms.expectNoteUpdatedToast()
       const noApplyFilters = false
 
       // Update note only gets exported
-      await adminPrograms.editNote('Note is updated')
+      await adminPrograms.editNoteForBulkStatus('Note is updated')
       await adminPrograms.expectNoteUpdatedToast()
+      await page.getByRole('link', {name: 'Back'}).click()
       const csvContent = await adminPrograms.getCsv(noApplyFilters)
       expect(csvContent).toContain('Note is updated')
 
-      await adminPrograms.viewApplicationForApplicant('Guest')
-      await adminPrograms.editNote('Note is finalized')
-      await adminPrograms.expectNoteUpdatedToast()
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
+      await adminPrograms.editNoteForBulkStatus('Note is finalized')
+      await page.getByRole('link', {name: 'Back'}).click()
       const csvContentFinal = await adminPrograms.getCsv(noApplyFilters)
       expect(csvContentFinal).toContain('Note is finalized')
     })
 
     test('preserves newlines in notes', async ({adminPrograms}) => {
+      await adminPrograms.viewApplications(programWithStatusesName)
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
       const noteText = 'Some note content\nwithseparatelines'
-      await adminPrograms.editNote(noteText)
+      await adminPrograms.editNoteForBulkStatus(noteText)
       await adminPrograms.expectNoteUpdatedToast()
 
-      // Reload the note editor.
-      await adminPrograms.viewApplications(programWithStatusesName)
-      await adminPrograms.viewApplicationForApplicant('Guest')
-
-      expect(await adminPrograms.getNoteContent()).toBe(noteText)
+      expect(await adminPrograms.getNoteContentForBulkStatus()).toBe(noteText)
     })
   })
 
@@ -397,6 +455,7 @@ test.describe('view program statuses', () => {
         applicantQuestions,
         adminProgramStatuses,
       }) => {
+        await enableFeatureFlag(page, 'bulk_status_update_enabled')
         await loginAsAdmin(page)
 
         // Add a program, no questions are needed.
@@ -441,31 +500,37 @@ test.describe('view program statuses', () => {
       page,
       adminPrograms,
     }) => {
-      await adminPrograms.expectApplicationHasStatusString(
+      await adminPrograms.expectApplicationHasStatusStringForBulkStatus(
         'Guest',
         `${waitingStatus} (default)`,
       )
-      await adminPrograms.expectApplicationHasStatusString(
+      await adminPrograms.expectApplicationHasStatusStringForBulkStatus(
         testUserDisplayName(),
         `${waitingStatus} (default)`,
       )
 
       // Approve guest application
-      await adminPrograms.viewApplicationForApplicant('Guest')
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
       const modal =
-        await adminPrograms.setStatusOptionAndAwaitModal(approvedStatus)
+        await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+          approvedStatus,
+        )
       expect(await modal.innerText()).toContain(
         `Status Change: ${waitingStatus} -> ${approvedStatus}`,
       )
-      await adminPrograms.confirmStatusUpdateModal(modal)
-      expect(await adminPrograms.getStatusOption()).toBe(approvedStatus)
+      await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+      expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+        approvedStatus,
+      )
       await adminPrograms.expectUpdateStatusToast()
 
-      await adminPrograms.expectApplicationStatusDoesntContain(
+      await page.getByRole('link', {name: 'Back'}).click()
+
+      await adminPrograms.expectApplicationStatusDoesntContainForBulkStatus(
         'Guest',
         '(default)',
       )
-      await adminPrograms.expectApplicationHasStatusString(
+      await adminPrograms.expectApplicationHasStatusStringForBulkStatus(
         testUserDisplayName(),
         `${waitingStatus} (default)`,
       )
@@ -489,6 +554,7 @@ test.describe('view program statuses', () => {
         applicantQuestions,
         adminProgramStatuses,
       }) => {
+        await enableFeatureFlag(page, 'bulk_status_update_enabled')
         await loginAsAdmin(page)
 
         // Add a program with a single question that is used for asserting downloaded content.
@@ -524,31 +590,31 @@ test.describe('view program statuses', () => {
     }) => {
       await adminPrograms.viewApplications(programForFilteringName)
       // Default page shows all applications.
-      await adminPrograms.expectApplicationCount(1)
+      await adminPrograms.expectApplicationCountForBulkStatus(1)
 
       // Included when filtering to applications without statuses.
       await adminPrograms.filterProgramApplications({
         applicationStatusOption:
           AdminPrograms.NO_STATUS_APPLICATION_FILTER_OPTION,
       })
-      await adminPrograms.expectApplicationCount(1)
+      await adminPrograms.expectApplicationCountForBulkStatus(1)
 
       // Excluded when selecting specific statuses.
       await adminPrograms.filterProgramApplications({
         applicationStatusOption: approvedStatusName,
       })
-      await adminPrograms.expectApplicationCount(0)
+      await adminPrograms.expectApplicationCountForBulkStatus(0)
       await adminPrograms.filterProgramApplications({
         applicationStatusOption: rejectedStatusName,
       })
-      await adminPrograms.expectApplicationCount(0)
+      await adminPrograms.expectApplicationCountForBulkStatus(0)
 
       // Included when explicitly selecting the default option to show all applications.
       await adminPrograms.filterProgramApplications({
         applicationStatusOption:
           AdminPrograms.ANY_STATUS_APPLICATION_FILTER_OPTION,
       })
-      await adminPrograms.expectApplicationCount(1)
+      await adminPrograms.expectApplicationCountForBulkStatus(1)
     })
 
     test('applied application status filter is used when downloading', async ({
@@ -587,14 +653,18 @@ test.describe('view program statuses', () => {
     })
 
     test('application with status shows in default filter and status-specific filter', async ({
+      page,
       adminPrograms,
     }) => {
       await test.step('explicitly set a status for the application', async () => {
         await adminPrograms.viewApplications(programForFilteringName)
-        await adminPrograms.viewApplicationForApplicant('Guest')
+        await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(approvedStatusName)
-        await adminPrograms.confirmStatusUpdateModal(modal)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            approvedStatusName,
+          )
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+        await page.getByRole('link', {name: 'Back'}).click()
       })
 
       // Excluded when filtering to applications without statuses.
@@ -602,53 +672,68 @@ test.describe('view program statuses', () => {
         applicationStatusOption:
           AdminPrograms.NO_STATUS_APPLICATION_FILTER_OPTION,
       })
-      await adminPrograms.expectApplicationCount(0)
+      await adminPrograms.expectApplicationCountForBulkStatus(0)
 
       // Included when selecting the "approved" status.
       await adminPrograms.filterProgramApplications({
         applicationStatusOption: approvedStatusName,
       })
-      await adminPrograms.expectApplicationCount(1)
+      await adminPrograms.expectApplicationCountForBulkStatus(1)
 
       // Excluded when selecting the "rejected" status.
       await adminPrograms.filterProgramApplications({
         applicationStatusOption: rejectedStatusName,
       })
-      await adminPrograms.expectApplicationCount(0)
+      await adminPrograms.expectApplicationCountForBulkStatus(0)
 
       // Included when explicitly selecting the default option to show all applications.
       await adminPrograms.filterProgramApplications({
         applicationStatusOption:
           AdminPrograms.ANY_STATUS_APPLICATION_FILTER_OPTION,
       })
-      await adminPrograms.expectApplicationCount(1)
+      await adminPrograms.expectApplicationCountForBulkStatus(1)
     })
 
     test('shows the application on reload after the status is updated to something no longer in the filter', async ({
+      page,
       adminPrograms,
     }) => {
       await test.step('explicitly set a status for the application', async () => {
         await adminPrograms.viewApplications(programForFilteringName)
-        await adminPrograms.viewApplicationForApplicant('Guest')
+        await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(approvedStatusName)
-        await adminPrograms.confirmStatusUpdateModal(modal)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            approvedStatusName,
+          )
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+        await page.getByRole('link', {name: 'Back'}).click()
       })
 
       await adminPrograms.filterProgramApplications({
         applicationStatusOption: approvedStatusName,
       })
-      await adminPrograms.viewApplicationForApplicant('Guest')
+      await adminPrograms.expectApplicationCountForBulkStatus(1)
+
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
       const modal =
-        await adminPrograms.setStatusOptionAndAwaitModal(rejectedStatusName)
-      await adminPrograms.confirmStatusUpdateModal(modal)
+        await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+          rejectedStatusName,
+        )
+      await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+      await page.getByRole('link', {name: 'Back'}).click()
 
       // The application should no longer be in the list, since its status is no longer "approved".
       // However, it should still be displayed in the viewer since admins may want to easily revert
       // the status update.
-      await adminPrograms.expectApplicationCount(0)
-      const applicationText = await adminPrograms
-        .applicationFrameLocator()
+      await adminPrograms.filterProgramApplications({
+        applicationStatusOption: approvedStatusName,
+      })
+      await adminPrograms.expectApplicationCountForBulkStatus(0)
+      await adminPrograms.filterProgramApplications({
+        applicationStatusOption: rejectedStatusName,
+      })
+      await adminPrograms.viewApplicationForApplicantForBulkStatus('Guest')
+      const applicationText = await page
         .locator('#application-view')
         .innerText()
       expect(applicationText).toContain('Guest')
@@ -668,9 +753,13 @@ test.describe('view program statuses', () => {
         applicantQuestions,
         adminPrograms,
       }) => {
+        await enableFeatureFlag(page, 'bulk_status_update_enabled')
         await loginAsAdmin(page)
 
         // Create a program without eligibility
+        await adminQuestions.addNameQuestion({
+          questionName: 'NameQuestion',
+        })
         await adminQuestions.addNumberQuestion({
           questionName: eligibilityQuestionId,
         })
@@ -681,7 +770,7 @@ test.describe('view program statuses', () => {
         await adminPrograms.editProgramBlock(
           eligibilityProgramName,
           'first description',
-          [eligibilityQuestionId, 'statuses-fave-color-q'],
+          [eligibilityQuestionId, 'statuses-fave-color-q', 'NameQuestion'],
         )
         await adminPrograms.gotoAdminProgramsPage()
         await adminPrograms.publishProgram(eligibilityProgramName)
@@ -693,6 +782,7 @@ test.describe('view program statuses', () => {
         // Fill out application and submit.
         await applicantQuestions.answerNumberQuestion('1')
         await applicantQuestions.answerTextQuestion('Red')
+        await applicantQuestions.answerNameQuestion('Robin', 'Hood')
         await applicantQuestions.clickNext()
         await applicantQuestions.submitFromReviewPage()
         await logout(page)
@@ -719,6 +809,7 @@ test.describe('view program statuses', () => {
         await applicantQuestions.applyProgram(eligibilityProgramName)
         await applicantQuestions.answerNumberQuestion('5')
         await applicantQuestions.answerTextQuestion('Red')
+        await applicantQuestions.answerNameQuestion('Sonny', 'Hood')
         await applicantQuestions.clickNext()
         await applicantQuestions.submitFromReviewPage()
 
@@ -726,16 +817,15 @@ test.describe('view program statuses', () => {
       },
     )
 
-    test('application list shows eligibility statuses', async ({
+    test('application table shows eligibility statuses', async ({
       page,
       adminPrograms,
     }) => {
       await loginAsProgramAdmin(page)
       await adminPrograms.viewApplications(eligibilityProgramName)
-      await adminPrograms.viewApplicationForApplicant('Guest')
       await validateScreenshot(
         page,
-        'application-view-with-eligibility-statuses',
+        'application-table-view-with-eligibility-statuses',
       )
     })
   })
@@ -754,6 +844,7 @@ test.describe('view program statuses', () => {
           applicantQuestions,
           adminProgramStatuses,
         )
+        await enableFeatureFlag(page, 'bulk_status_update_enabled')
         await loginAsAdmin(page)
         await adminQuestions.addEmailQuestion({
           questionName: 'Email',
@@ -782,13 +873,13 @@ test.describe('view program statuses', () => {
         await applicantQuestions.answerEmailQuestion(guestEmail)
         await applicantQuestions.clickNext()
         await applicantQuestions.submitFromReviewPage()
+        const id = await adminPrograms.getApplicationId()
         await logout(page)
-      })
-
-      await test.step('view submitted programs as a program admin', async () => {
         await loginAsProgramAdmin(page)
         await adminPrograms.viewApplications(programWithStatusesName)
-        await adminPrograms.viewApplicationForApplicant('Guest')
+        await adminPrograms.viewApplicationForApplicantForBulkStatus(
+          `${guestEmail} (${id})`,
+        )
       })
 
       const emailsBefore =
@@ -800,7 +891,9 @@ test.describe('view program statuses', () => {
 
       await test.step('set new status and confirm change via modal', async () => {
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            emailStatusName,
+          )
 
         const notifyCheckbox = await modal.$('input[type=checkbox]')
         expect(notifyCheckbox).not.toBeNull()
@@ -808,8 +901,10 @@ test.describe('view program statuses', () => {
         expect(await modal.innerText()).toContain(
           ' of this change at ' + guestEmail,
         )
-        await adminPrograms.confirmStatusUpdateModal(modal)
-        expect(await adminPrograms.getStatusOption()).toBe(emailStatusName)
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+        expect(await adminPrograms.getStatusOptionForBulkStatus()).toBe(
+          emailStatusName,
+        )
         await adminPrograms.expectUpdateStatusToast()
       })
       await test.step('verify status update email was sent to applicant', async () => {
@@ -836,13 +931,13 @@ test.describe('view program statuses', () => {
         await applicantQuestions.answerEmailQuestion(otherTestUserEmail)
         await applicantQuestions.clickNext()
         await applicantQuestions.submitFromReviewPage()
+        const id = await adminPrograms.getApplicationId()
         await logout(page)
-      })
-
-      await test.step('view submitted programs as a program admin', async () => {
         await loginAsProgramAdmin(page)
         await adminPrograms.viewApplications(programWithStatusesName)
-        await adminPrograms.viewApplicationForApplicant(otherTestUserEmail)
+        await adminPrograms.viewApplicationForApplicantForBulkStatus(
+          `${otherTestUserEmail} (${id})`,
+        )
       })
 
       const [acccountEmailsBefore, applicantEmailsBefore] =
@@ -858,7 +953,9 @@ test.describe('view program statuses', () => {
 
       await test.step('set new status and confirm change via modal', async () => {
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            emailStatusName,
+          )
         expect(await modal.innerText()).toContain(
           ' of this change at ' +
             testUserDisplayName() +
@@ -866,7 +963,7 @@ test.describe('view program statuses', () => {
             otherTestUserEmail,
         )
 
-        await adminPrograms.confirmStatusUpdateModal(modal)
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
         await adminPrograms.expectUpdateStatusToast()
       })
       await test.step('verify status update email was sent to both email addresses', async () => {
@@ -897,13 +994,13 @@ test.describe('view program statuses', () => {
         await applicantQuestions.answerEmailQuestion(testUserDisplayName())
         await applicantQuestions.clickNext()
         await applicantQuestions.submitFromReviewPage()
+        const id = await adminPrograms.getApplicationId()
         await logout(page)
-      })
-
-      await test.step('view submitted programs as a program admin', async () => {
         await loginAsProgramAdmin(page)
         await adminPrograms.viewApplications(programWithStatusesName)
-        await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
+        await adminPrograms.viewApplicationForApplicantForBulkStatus(
+          `${testUserDisplayName()} (${id})`,
+        )
       })
 
       const emailsBefore =
@@ -915,13 +1012,15 @@ test.describe('view program statuses', () => {
 
       await test.step('set new status and confirm change via modal', async () => {
         const modal =
-          await adminPrograms.setStatusOptionAndAwaitModal(emailStatusName)
+          await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+            emailStatusName,
+          )
         expect(await modal.innerText()).toContain(
           ' of this change at ' + testUserDisplayName(),
         )
         expect(await modal.innerText()).not.toContain(' and ')
 
-        await adminPrograms.confirmStatusUpdateModal(modal)
+        await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
         await adminPrograms.expectUpdateStatusToast()
       })
       await test.step('verify status update email was sent to the applicant', async () => {
