+++ browser-test/src/admin/admin_predicates.test.ts
@@ -21,6 +21,7 @@ test.describe('create and edit predicates', () => {
     adminPredicates,
   }) => {
     await loginAsAdmin(page)
+    await enableFeatureFlag(page, 'bulk_status_update_enabled')
 
     // Add a program with two screens
     await adminQuestions.addTextQuestion({questionName: 'hide-predicate-q'})
@@ -102,13 +103,13 @@ test.describe('create and edit predicates', () => {
     await logout(page)
     await loginAsProgramAdmin(page)
     await adminPrograms.viewApplications(programName)
-    await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
+    await adminPrograms.viewApplicationForApplicantForBulkStatus(
+      testUserDisplayName(),
+    )
 
-    const applicationText = await adminPrograms
-      .applicationFrameLocator()
-      .locator('#application-view')
-      .innerText()
-    expect(applicationText).not.toContain('Screen 2')
+    expect(await page.innerHTML('#application-view')).not.toContain('Screen 2')
+
+    await page.getByRole('link', {name: 'Back'}).click()
   })
 
   test('add a show predicate', async ({
@@ -119,6 +120,7 @@ test.describe('create and edit predicates', () => {
     adminPredicates,
   }) => {
     await loginAsAdmin(page)
+    await enableFeatureFlag(page, 'bulk_status_update_enabled')
 
     // Add a program with two screens
     await adminQuestions.addTextQuestion({
@@ -206,13 +208,14 @@ test.describe('create and edit predicates', () => {
     await loginAsProgramAdmin(page)
     await adminPrograms.viewApplications(programName)
 
-    await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
-    expect(
-      await adminPrograms
-        .applicationFrameLocator()
-        .locator('#application-view')
-        .innerText(),
-    ).toContain('Screen 2')
+    await adminPrograms.viewApplicationForApplicantForBulkStatus(
+      testUserDisplayName(),
+    )
+    expect(await page.locator('#application-view').innerText()).toContain(
+      'Screen 2',
+    )
+
+    await page.getByRole('link', {name: 'Back'}).click()
   })
 
   test('add an eligibility predicate', async ({
@@ -224,6 +227,7 @@ test.describe('create and edit predicates', () => {
   }) => {
     await loginAsAdmin(page)
 
+    await enableFeatureFlag(page, 'bulk_status_update_enabled')
     // Add a program with two screens
     await adminQuestions.addTextQuestion({
       questionName: 'eligibility-predicate-q',
@@ -350,13 +354,14 @@ test.describe('create and edit predicates', () => {
     await loginAsProgramAdmin(page)
     await adminPrograms.viewApplications(programName)
 
-    await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
-    expect(
-      await adminPrograms
-        .applicationFrameLocator()
-        .locator('#application-view')
-        .innerText(),
-    ).toContain('Screen 1')
+    await adminPrograms.viewApplicationForApplicantForBulkStatus(
+      testUserDisplayName(),
+    )
+    expect(await page.locator('#application-view').innerText()).toContain(
+      'Screen 1',
+    )
+
+    await page.getByRole('link', {name: 'Back'}).click()
   })
 
   test('suffix cannot be added as an eligibility predicate for name question', async ({
@@ -398,7 +403,6 @@ test.describe('create and edit predicates', () => {
       await page.getByText('name suffix').isHidden()
     })
   })
-
   test('eligibility message field is available to use', async ({
     page,
     adminQuestions,
