+++ browser-test/src/applicant/applicant_application_statuses.test.ts
@@ -7,6 +7,7 @@ import {
   testUserDisplayName,
   validateAccessibility,
   validateScreenshot,
+  enableFeatureFlag,
 } from '../support'
 
 test.describe('with program statuses', () => {
@@ -15,6 +16,7 @@ test.describe('with program statuses', () => {
 
   test.beforeEach(
     async ({page, adminPrograms, adminProgramStatuses, applicantQuestions}) => {
+      await enableFeatureFlag(page, 'bulk_status_update_enabled')
       await loginAsAdmin(page)
 
       await adminPrograms.addProgram(programName)
@@ -33,10 +35,15 @@ test.describe('with program statuses', () => {
       // Navigate to the submitted application as the program admin and set a status.
       await loginAsProgramAdmin(page)
       await adminPrograms.viewApplications(programName)
-      await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
+      await adminPrograms.viewApplicationForApplicantForBulkStatus(
+        testUserDisplayName(),
+      )
       const modal =
-        await adminPrograms.setStatusOptionAndAwaitModal(approvedStatusName)
-      await adminPrograms.confirmStatusUpdateModal(modal)
+        await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+          approvedStatusName,
+        )
+      await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+      await page.getByRole('link', {name: 'Back'}).click()
       await logout(page)
     },
   )
