+++ browser-test/src/applicant/application_version_fast_forward.test.ts
@@ -11,6 +11,7 @@ import {
   AdminPredicates,
   enableFeatureFlag,
   disableFeatureFlag,
+  testUserDisplayName,
 } from '../support'
 import {ProgramVisibility, QuestionSpec} from '../support/admin_programs'
 import {Browser, Locator, Page} from '@playwright/test'
@@ -37,7 +38,10 @@ test.describe('Application Version Fast-Forward Flow', () => {
       programName,
       browser,
     )
-
+    await enableFeatureFlag(
+      civiformAdminActor.getPage(),
+      'BULK_STATUS_UPDATE_ENABLED',
+    )
     await disableFeatureFlag(
       civiformAdminActor.getPage(),
       'FASTFORWARD_ENABLED',
@@ -463,27 +467,27 @@ test.describe('Application Version Fast-Forward Flow', () => {
       })
 
       await test.step('sees submitted application with questions from program version v1', async () => {
-        const cardLocator = programAdminActor.getCardLocator()
-        await expect(cardLocator).toHaveCount(1)
+        const rowLocator = programAdminActor.getRowLocator()
+        await expect(rowLocator).toHaveCount(1)
 
-        const cardButton = cardLocator.getByRole('link', {name: 'View'})
-        expect(
-          await programAdminActor.parseProgramIdFromLocator(cardButton),
-        ).toBe(programIdV1)
+        const row = rowLocator.getByRole('link', {name: testUserDisplayName()})
+        expect(await programAdminActor.parseProgramIdFromLocator(row)).toBe(
+          programIdV1,
+        )
       })
 
       await test.step('does not see submitted application with questions from program version v2 or v3', async () => {
-        const cardLocator = programAdminActor.getCardLocator()
-        await expect(cardLocator).toHaveCount(1)
+        const rowLocator = programAdminActor.getRowLocator()
+        await expect(rowLocator).toHaveCount(1)
 
-        const cardButton = cardLocator.getByRole('link', {name: 'View'})
-        expect(
-          await programAdminActor.parseProgramIdFromLocator(cardButton),
-        ).not.toBe(programIdV2)
+        const row = rowLocator.getByRole('link', {name: testUserDisplayName()})
+        expect(await programAdminActor.parseProgramIdFromLocator(row)).not.toBe(
+          programIdV2,
+        )
 
-        expect(
-          await programAdminActor.parseProgramIdFromLocator(cardButton),
-        ).not.toBe(programIdV3)
+        expect(await programAdminActor.parseProgramIdFromLocator(row)).not.toBe(
+          programIdV3,
+        )
       })
     })
   })
@@ -503,7 +507,10 @@ test.describe('Application Version Fast-Forward Flow', () => {
       programName,
       browser,
     )
-
+    await enableFeatureFlag(
+      civiformAdminActor.getPage(),
+      'BULK_STATUS_UPDATE_ENABLED',
+    )
     await enableFeatureFlag(civiformAdminActor.getPage(), 'FASTFORWARD_ENABLED')
     await enableFeatureFlag(applicantActor.getPage(), 'FASTFORWARD_ENABLED')
     await enableFeatureFlag(programAdminActor.getPage(), 'FASTFORWARD_ENABLED')
@@ -956,27 +963,27 @@ test.describe('Application Version Fast-Forward Flow', () => {
       })
 
       await test.step('sees submitted application with questions from program version v3', async () => {
-        const cardLocator = programAdminActor.getCardLocator()
-        await expect(cardLocator).toHaveCount(1)
+        const rowLocator = programAdminActor.getRowLocator()
+        await expect(rowLocator).toHaveCount(1)
 
-        const cardButton = cardLocator.getByRole('link', {name: 'View'})
-        expect(
-          await programAdminActor.parseProgramIdFromLocator(cardButton),
-        ).toBe(programIdV3)
+        const row = rowLocator.getByRole('link', {name: testUserDisplayName()})
+        expect(await programAdminActor.parseProgramIdFromLocator(row)).toBe(
+          programIdV3,
+        )
       })
 
       await test.step('does not see submitted application with questions from program version v1 or v2', async () => {
-        const cardLocator = programAdminActor.getCardLocator()
-        await expect(cardLocator).toHaveCount(1)
+        const rowLocator = programAdminActor.getRowLocator()
+        await expect(rowLocator).toHaveCount(1)
 
-        const cardButton = cardLocator.getByRole('link', {name: 'View'})
-        expect(
-          await programAdminActor.parseProgramIdFromLocator(cardButton),
-        ).not.toBe(programIdV1)
+        const row = rowLocator.getByRole('link', {name: testUserDisplayName()})
+        expect(await programAdminActor.parseProgramIdFromLocator(row)).not.toBe(
+          programIdV1,
+        )
 
-        expect(
-          await programAdminActor.parseProgramIdFromLocator(cardButton),
-        ).not.toBe(programIdV2)
+        expect(await programAdminActor.parseProgramIdFromLocator(row)).not.toBe(
+          programIdV2,
+        )
       })
     })
 
@@ -998,7 +1005,10 @@ test.describe('Application Version Fast-Forward Flow', () => {
       programName,
       browser,
     )
-
+    await enableFeatureFlag(
+      civiformAdminActor.getPage(),
+      'BULK_STATUS_UPDATE_ENABLED',
+    )
     await enableFeatureFlag(civiformAdminActor.getPage(), 'FASTFORWARD_ENABLED')
     await enableFeatureFlag(applicantActor.getPage(), 'FASTFORWARD_ENABLED')
 
@@ -1982,11 +1992,11 @@ class FastForwardProgramAdminActor {
   }
 
   /**
-   * Get a locator any admin application cards. There may be 0-♾️
-   * @returns {Locator} Locator to admin application card
+   * Get a locator any admin application rows. There may be 0-♾️
+   * @returns {Locator} Locator to admin application row
    */
-  getCardLocator(): Locator {
-    return this.page.locator('.cf-admin-application-card')
+  getRowLocator(): Locator {
+    return this.page.locator('.cf-admin-application-row')
   }
 
   /**
