+++ browser-test/src/applicant/northstar_applicant_application_statuses.test.ts
@@ -8,7 +8,6 @@ import {
   testUserDisplayName,
   validateAccessibility,
   validateScreenshot,
-  normalizeElements,
 } from '../support'
 
 test.describe('with program statuses', {tag: ['@northstar']}, () => {
@@ -17,6 +16,7 @@ test.describe('with program statuses', {tag: ['@northstar']}, () => {
 
   test.beforeEach(
     async ({page, adminPrograms, adminProgramStatuses, applicantQuestions}) => {
+      await enableFeatureFlag(page, 'bulk_status_update_enabled')
       await loginAsAdmin(page)
 
       await adminPrograms.addProgram(programName)
@@ -35,10 +35,15 @@ test.describe('with program statuses', {tag: ['@northstar']}, () => {
       // Navigate to the submitted application as the program admin and set a status.
       await loginAsProgramAdmin(page)
       await adminPrograms.viewApplications(programName)
-      await adminPrograms.viewApplicationForApplicant(testUserDisplayName())
+      await adminPrograms.viewApplicationForApplicantForBulkStatus(
+        testUserDisplayName(),
+      )
       const modal =
-        await adminPrograms.setStatusOptionAndAwaitModal(approvedStatusName)
-      await adminPrograms.confirmStatusUpdateModal(modal)
+        await adminPrograms.setStatusOptionAndAwaitModalForBulkStatus(
+          approvedStatusName,
+        )
+      await adminPrograms.confirmStatusUpdateModalForBulkStatus(modal)
+      await page.getByRole('link', {name: 'Back'}).click()
       await logout(page)
 
       await enableFeatureFlag(page, 'north_star_applicant_ui')
@@ -52,11 +57,9 @@ test.describe('with program statuses', {tag: ['@northstar']}, () => {
       await loginAsTestUser(page)
 
       const locator = page.locator('.cf-application-card')
-      await normalizeElements(page)
+
       await expect(locator.getByText('Submitted on 1/1/30')).toBeHidden()
-      await expect(
-        locator.getByText(approvedStatusName + ' on 1/1/30'),
-      ).toBeVisible()
+      await expect(locator.getByText(approvedStatusName)).toBeVisible()
 
       await validateScreenshot(locator, 'program-card-with-status-northstar')
       await validateAccessibility(page)
