+++ browser-test/src/primary_applicant_info.test.ts
@@ -6,6 +6,7 @@ import {
   logout,
   validateScreenshot,
   waitForPageJsLoad,
+  enableFeatureFlag,
 } from './support'
 import {
   PrimaryApplicantInfoAlertType,
@@ -234,6 +235,7 @@ test.describe('primary applicant info questions', () => {
     applicantQuestions,
   }) => {
     await loginAsAdmin(page)
+    await enableFeatureFlag(page, 'bulk_status_update_enabled')
 
     await adminQuestions.addNameQuestion({
       questionName: 'name',
@@ -260,9 +262,7 @@ test.describe('primary applicant info questions', () => {
     await loginAsProgramAdmin(page)
 
     await adminPrograms.viewApplications('test')
-    await expect(
-      page.locator(adminPrograms.selectApplicationCardForApplicant('LaForge')),
-    ).toBeVisible()
+    expect(adminPrograms.getRowLocator('LaForge')).not.toBeNull()
 
     await logout(page)
     await loginAsTestUser(
@@ -275,8 +275,6 @@ test.describe('primary applicant info questions', () => {
     await loginAsProgramAdmin(page)
 
     await adminPrograms.viewApplications('test')
-    await expect(
-      page.locator(adminPrograms.selectApplicationCardForApplicant('LaForge')),
-    ).toBeVisible()
+    expect(adminPrograms.getRowLocator('LaForge')).not.toBeNull()
   })
 })
